refactor: simplify main Sample
