--- samples/KafkaFlow.Sample/TestMessage.cs
@@ -1,11 +1,0 @@
-﻿namespace KafkaFlow.Sample
-{
-    using System.Runtime.Serialization;
-
-    [DataContract]
-    public class TestMessage
-    {
-        [DataMember(Order = 1)]
-        public string Text { get; set; }
-    }
-}
